/home/hjalte/Desktop/cp-contest/twosum.cpp finished and accepted!
